+++ src/Document/ResourceObject.php
@@ -17,8 +17,6 @@
 
 namespace CloudCreativity\LaravelJsonApi\Document;
 
-use CloudCreativity\LaravelJsonApi\Contracts\Object\ResourceObjectInterface;
-use CloudCreativity\LaravelJsonApi\Object\ResourceObject as LegacyResourceObject;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
@@ -614,19 +612,6 @@ class ResourceObject implements Arrayable, \IteratorAggregate, \JsonSerializable
     public function jsonSerialize()
     {
         return $this->toArray();
-    }
-
-    /**
-     * Convert the object to a legacy resource object.
-     *
-     * @return ResourceObjectInterface
-     * @deprecated 2.0.0
-     */
-    public function toObject(): ResourceObjectInterface
-    {
-        $json = json_encode($this);
-
-        return new LegacyResourceObject(json_decode($json, false));
     }
 
     /**