+++ src/Http/Requests/ValidatedRequest.php
@@ -19,14 +19,12 @@ namespace CloudCreativity\LaravelJsonApi\Http\Requests;
 
 use CloudCreativity\LaravelJsonApi\Contracts\Auth\AuthorizerInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\ContainerInterface;
-use CloudCreativity\LaravelJsonApi\Contracts\Object\DocumentInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Validation\DocumentValidatorInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Validation\ValidatorFactoryInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Validation\ValidatorInterface;
 use CloudCreativity\LaravelJsonApi\Exceptions\DocumentRequiredException;
 use CloudCreativity\LaravelJsonApi\Exceptions\ValidationException;
 use CloudCreativity\LaravelJsonApi\Factories\Factory;
-use CloudCreativity\LaravelJsonApi\Object\Document;
 use CloudCreativity\LaravelJsonApi\Routing\Route;
 use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Auth\AuthenticationException;
@@ -37,6 +35,11 @@ use Illuminate\Support\Arr;
 use Neomerx\JsonApi\Contracts\Encoder\Parameters\EncodingParametersInterface;
 use Neomerx\JsonApi\Exceptions\JsonApiException;
 
+/**
+ * Class ValidatedRequest
+ *
+ * @package CloudCreativity\LaravelJsonApi
+ */
 abstract class ValidatedRequest implements ValidatesWhenResolved
 {
 
@@ -198,21 +201,6 @@ abstract class ValidatedRequest implements ValidatesWhenResolved
     public function getResourceType()
     {
         return $this->route->getResourceType();
-    }
-
-    /**
-     * Get the validated JSON API document, if there is one.
-     *
-     * @return DocumentInterface|null
-     * @deprecated 2.0.0
-     */
-    public function getDocument()
-    {
-        if (!$document = $this->decode()) {
-            return null;
-        }
-
-        return new Document($document);
     }
 
     /**